requirements.txt
Flask==2.3.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.1
matplotlib==3.7.2
paho-mqtt==1.6.1
